// projects-roadmap-sh/backend/proj-02-github-user-activity/src/types/github.types.ts
export interface GithubEventType {
  id?: string;
  type?: string;
  actor?: Actor;
  repo?: WelcomeRepo;
  payload?: Payload;
  public?: boolean;
  created_at?: string;
}

export interface Actor {
  id?: number;
  login?: string;
  display_login?: string;
  gravatar_id?: string;
  url?: string;
  avatar_url?: string;
}

export interface Payload {
  repository_id?: number;
  push_id?: number;
  size?: number;
  distinct_size?: number;
  ref?: null | string;
  head?: string;
  before?: string;
  commits?: Commit[];
  ref_type?: string;
  pusher_type?: string;
  action?: string;
  number?: number;
  pull_request?: PullRequest;
  master_branch?: string;
  description?: null | string;
}

export interface Commit {
  sha: string;
  author: Author;
  message: string;
  distinct: boolean;
  url: string;
}

export interface Author {
  email: string;
  name: string;
}

export interface PullRequest {
  url: string;
  id: number;
  node_id: string;
  html_url: string;
  diff_url: string;
  patch_url: string;
  issue_url: string;
  number: number;
  state: string;
  locked: boolean;
  title: string;
  user: User;
  body: null;
  created_at: Date;
  updated_at: Date;
  closed_at: null;
  merged_at: null;
  merge_commit_sha: null;
  assignee: null;
  assignees: any[];
  requested_reviewers: any[];
  requested_teams: any[];
  labels: any[];
  milestone: null;
  draft: boolean;
  commits_url: string;
  review_comments_url: string;
  review_comment_url: string;
  comments_url: string;
  statuses_url: string;
  head: Base;
  base: Base;
  _links: Links;
  author_association: string;
  auto_merge: null;
  active_lock_reason: null;
  merged: boolean;
  mergeable: null;
  rebaseable: null;
  mergeable_state: string;
  merged_by: null;
  comments: number;
  review_comments: number;
  maintainer_can_modify: boolean;
  commits: number;
  additions: number;
  deletions: number;
  changed_files: number;
}
// ...
export interface WelcomeRepo {
  id?: number;
  name?: string;
  url?: string;
}